move macro stuff around
